+++ LearnOpenGLTuto/shaders/shader_texture_phong.frag
@@ -17,7 +17,7 @@ uniform vec3 viewPosition;
 uniform float ambientStrength;
 uniform float diffuseStrength;
 uniform float specularStrength;
-uniform int shininess;
+uniform float shininess;
 
 void main()
 {
